provisional (alpha) version of 4.7 code
