+++ info.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
-<extension key="uk.co.circleinteractive.payment.sagepay" type="payment">
-  <downloadUrl>http://extensions.circle-interactive.co.uk/uk.co.circleinteractive.payment.sagepay-v3.1.0.zip</downloadUrl>
+<extension key="uk.co.circleinteractive.payment.sagepay" type="module">
+  <downloadUrl>http://extensions.circle-interactive.co.uk/uk.co.circleinteractive.payment.sagepay-v4.0.0.zip</downloadUrl>
   <file>sagepay</file>
   <name>Sagepay</name>
   <description>Sagepay Payment Processor</description>
@@ -12,8 +12,8 @@
     <author>Andrew Walker - Circle Interactive</author>
     <email>andyw (at) circle-interactive (dot) co (dot) uk</email>
   </maintainer>
-  <releaseDate>2016-01-12</releaseDate>
-  <version>3.1.0</version>
+  <releaseDate>2016-08-01</releaseDate>
+  <version>4.0.0</version>
   <develStage>stable</develStage>
   <compatibility>
     <ver>4.2</ver>
@@ -21,23 +21,7 @@
     <ver>4.4</ver>
     <ver>4.5</ver>
     <ver>4.6</ver>
+    <ver>4.7</ver>
   </compatibility>
   <comments>For support, please contact project team on the forums.</comments>
-  <typeInfo>
-   <userNameLabel>Vendor Name</userNameLabel>
-   <passwordLabel></passwordLabel>
-   <signatureLabel></signatureLabel>
-   <subjectLabel></subjectLabel>
-   <urlSiteDefault>https://live.sagepay.com/gateway/service/vspserver-register.vsp</urlSiteDefault>
-   <urlApiDefault></urlApiDefault>
-   <urlRecurDefault>https://live.sagepay.com/gateway/service/repeat.vsp</urlRecurDefault>
-   <urlSiteTestDefault>https://test.sagepay.com/gateway/service/vspserver-register.vsp</urlSiteTestDefault>
-   <urlApiTestDefault></urlApiTestDefault>
-   <urlRecurTestDefault>https://test.sagepay.com/gateway/service/repeat.vsp</urlRecurTestDefault>
-   <urlButtonDefault></urlButtonDefault>
-   <urlButtonTestDefault></urlButtonTestDefault>
-   <billingMode>notify</billingMode>
-   <isRecur>1</isRecur>
-   <paymentType>1</paymentType>
-  </typeInfo>
 </extension>
